+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/req/ToInfo.java
@@ -4,9 +4,9 @@ import java.math.BigInteger;
 
 public class ToInfo {
   public String address;
-  public BigInteger amount;
+  public long amount;
 
-  public ToInfo(String address, BigInteger amount) {
+  public ToInfo(String address, long amount) {
     this.address = address;
     this.amount = amount;
   }
